better make install
darcs-hash:20041211000916-28bda-573ce1d7e07d7935267f0128e3663d3fe388d4bc.gz

+++ Makefile
@@ -5,7 +5,7 @@ test:
 install: 
 	install -d /etc/clog/filters
 	install lib/* /etc/clog/filters
-	install etc/clog.conf /etc/clog
+	[ -f /etc/clog/clog.conf ] || install etc/clog.conf /etc/clog
 	install -d /usr/local/bin
 	install bin/clog.rb /usr/local/bin/clog
 .PHONY: test install
